[fix] component 위치 수정, svg 위치 수정
