--- src/app/mypage/components/ProfileCard.tsx
@@ -1,94 +1,0 @@
-// import ProfileEditIcon from "../icons/profile-edit-icon.svg";
-
-export default function ProfileCard() {
-  return (
-    <div className="rounded-lg overflow-hidden border mb-4">
-      {/* Card Header with Orange Background */}
-      <div className="bg-main-450 p-4 relative">
-        <div className="text-gray-800 text-lg font-bold mb-1">내 프로필</div>
-        <div className="absolute right-4 top-4">
-          <button>
-            <svg width="32" height="32" viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <circle cx="16" cy="16" r="15" fill="#E5E7EB" stroke="#E5E7EB" strokeWidth="2"/>
-            <path d="M16.0208 9.22574C16.2067 8.82492 16.6817 8.65204 17.0818 8.83958L20.525 10.4538C20.9251 10.6414 21.0987 11.1183 20.9128 11.5191L16.1461 21.7962C16.0555 21.9913 15.8904 22.1415 15.6878 22.2128L13.1831 23.094C12.7735 23.2381 12.3233 23.027 12.1698 22.6189L11.2312 20.1235C11.1554 19.9217 11.1636 19.698 11.2541 19.5028L16.0208 9.22574Z" fill="#9CA3AF"/>
-            <path d="M12.667 9.56641L22.2024 14.0401" stroke="#E5E7EB"/>
-            </svg>
-          </button>
-        </div>
-        {/*Background */}
-        <div className="h-6 mt-2 relative">
-          <div className="absolute bottom-0 left-0 right-0 flex items-end justify-center">
-            <svg width="159" height="49" viewBox="0 0 159 49" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <rect x="40.209" y="20.8491" width="35.7907" height="3.97674" fill="#A065FF"/>
-            <rect x="40.209" y="24.8257" width="3.53488" height="22.093" fill="#A065FF"/>
-            <rect x="53.9072" y="24.8257" width="3.53488" height="22.093" fill="#A065FF"/>
-            <rect x="76" y="24.8257" width="3.53488" height="22.093" fill="#C3A7FF"/>
-            <rect x="91.4648" y="24.8257" width="3.53488" height="22.093" fill="#C3A7FF"/>
-            <rect x="76" y="20.8491" width="19" height="3.97674" fill="#C3A7FF"/>
-            <rect y="13.7793" width="16.3488" height="16.1279" fill="#A065FF"/>
-            <rect y="29.9072" width="16.3488" height="3.97674" fill="#C3A7FF"/>
-            <rect x="16.3486" y="29.9072" width="10.6047" height="3.97674" fill="#A065FF"/>
-            <rect x="12.8145" y="33.8838" width="3.53488" height="13.2558" fill="#A065FF"/>
-            <rect y="33.8838" width="3.53488" height="13.2558" fill="#A065FF"/>
-            <rect x="23.4189" y="33.8838" width="3.53488" height="13.2558" fill="#A065FF"/>
-            <path d="M58.2998 19.1108H64.2448H76.0485V20.8681H58.2998V19.1108Z" fill="#7E36E3"/>
-            <path d="M64.2744 19.1108L64.6259 10.5H76.6634L76.0483 19.1108H64.2744Z" fill="#C3A7FF"/>
-            <path d="M65.5928 18.0562L66.0321 11.6421H75.4336L74.9065 18.0562H65.5928Z" fill="#D3C1FF"/>
-            <path d="M115.481 30.5696C113.96 30.8855 112.713 31.568 111.693 32.6699C110.896 33.5282 110.539 34.2669 110.565 35.0412C110.59 35.7442 110.838 36.2239 111.344 36.571C112.221 37.1582 113.358 37.0821 114.201 36.3852L114.632 36.0211L114.707 36.4364C114.804 36.9222 115.144 37.4289 115.492 37.5884C116.128 37.8793 117.235 37.5493 117.727 36.9073C117.968 36.6017 117.968 36.6017 118.086 36.8011C118.183 36.9706 118.169 37.1548 118.015 37.6507C117.158 40.5723 113.472 43.5954 108.077 45.862C106.954 46.3253 104.08 47.9399 104.664 47.9165C105.087 47.8996 109.873 45.5978 111.505 44.7061C115.057 42.7683 117.886 40.0534 118.472 38.0303L118.625 37.5242L119.22 37.5107C119.572 37.4966 119.863 37.4441 119.94 37.3594C120.319 36.9463 120.303 37.3653 119.907 38.738C118.926 42.1237 116.918 44.9894 114.116 47.04C113.379 47.5796 113.16 47.5091 113.354 47.5829C113.518 47.6478 113.848 47.7241 114.321 47.4093C116.419 46.0092 118.539 43.4554 119.654 41.0131C120.037 40.1612 120.746 37.9291 120.735 37.603C120.727 37.3585 120.847 37.3537 121.088 37.5991C121.569 38.1002 122.544 38.0102 123.107 37.4061L123.428 37.0668L123.582 37.4177C123.669 37.6182 123.797 37.807 123.858 37.8351C124.002 37.9212 124.044 39.1234 123.946 40.0864C123.611 43.2115 122.067 46.079 119.741 47.8963C119.706 47.9169 121.101 47.8256 121.377 47.9166C121.377 47.9166 120.332 48.0767 121.257 47.1317C123.311 45.0294 124.465 42.2489 124.496 39.3706L124.515 38.1761L125.009 38.1564C125.562 38.1342 126.187 37.8338 126.416 37.4675C126.799 36.8808 126.353 36.7967 125.838 37.358C125.556 37.6652 125.427 37.7315 125.055 37.7464C124.692 37.7609 124.549 37.7054 124.299 37.4604C124.007 37.1762 123.995 37.1257 123.963 36.1986C123.93 35.2409 123.929 35.2307 123.689 35.2709C123.498 35.2989 123.422 35.4244 123.297 35.8987C123.137 36.5071 122.833 37.06 122.501 37.3589C121.963 37.8396 121.262 37.3881 121.119 36.4654C121.068 36.1614 121.051 35.6621 121.071 35.3553C121.09 35.0484 121.09 34.7627 121.068 34.7126C120.994 34.5932 120.633 34.6484 120.597 34.7825C120.579 34.8546 120.492 35.2356 120.415 35.6264C120.223 36.5013 119.731 37.1433 119.228 37.1635C118.765 37.182 118.561 36.8331 118.456 35.8272C118.391 35.1157 118.347 35.0052 118.196 35.0112C118.065 35.0165 117.959 35.1737 117.833 35.5869C117.398 37.0021 116.294 37.7298 115.585 37.0644C115.325 36.8198 115.312 36.7488 115.268 35.761C115.238 34.895 115.201 34.7128 115.07 34.6977C114.978 34.6809 114.831 34.799 114.726 34.9563C114.028 36.0657 113.193 36.6806 112.524 36.5748C111.662 36.4358 111.009 35.6254 111.102 34.8055C111.275 33.411 112.952 31.7727 114.866 31.1553C115.764 30.8644 119.106 30.7204 120.474 30.9309C122.287 31.195 123.593 31.6325 124.85 32.4086C126.487 33.4143 127.365 34.6034 127.716 36.2728C128.357 39.3793 126.644 44.3246 123.911 47.2193C123.396 47.7602 123.361 47.9147 123.743 47.8994C124.116 47.8845 125.412 46.3226 126.26 44.8807C128.44 41.1716 129.013 36.7616 127.62 34.3892C126.473 32.4252 123.894 31.0184 120.592 30.5589C119.316 30.3855 116.382 30.3907 115.481 30.5696Z" fill="#7E36E3" stroke="#7E36E3" strokeWidth="0.8"/>
-            <path d="M134.197 36.3957C133.137 36.6109 132.268 37.0743 131.556 37.8217C131 38.4038 130.751 38.9046 130.769 39.4293C130.786 39.9056 130.958 40.2305 131.311 40.4652C131.921 40.8625 132.713 40.81 133.301 40.3372L133.601 40.0901L133.653 40.3715C133.721 40.7006 133.957 41.0436 134.2 41.1515C134.642 41.348 135.414 41.1236 135.757 40.6882C135.926 40.481 135.926 40.481 136.008 40.616C136.075 40.7308 136.065 40.8556 135.957 41.1917C135.359 43.172 132.789 45.2232 129.028 46.7631C128.245 47.0779 126.242 48.1741 126.649 48.1578C126.943 48.146 130.279 46.5828 131.417 45.9773C133.893 44.6616 135.866 42.8199 136.275 41.4486L136.383 41.1056L136.797 41.0959C137.042 41.0861 137.245 41.0503 137.299 40.9929C137.563 40.7127 137.552 40.9966 137.275 41.927C136.589 44.2218 135.188 46.1651 133.235 47.5567C132.721 47.9229 132.569 47.8752 132.704 47.9251C132.818 47.969 133.047 48.0204 133.377 47.8067C134.84 46.8564 136.318 45.1244 137.097 43.4688C137.365 42.8912 137.86 41.3782 137.852 41.1573C137.847 40.9917 137.931 40.9883 138.098 41.1544C138.433 41.4936 139.112 41.4318 139.505 41.0221L139.729 40.7919L139.836 41.0296C139.897 41.1654 139.985 41.2932 140.028 41.3122C140.128 41.3704 140.157 42.185 140.088 42.8375C139.852 44.9553 138.775 46.8994 137.153 48.1326C137.128 48.1465 138.101 48.0836 138.293 48.1451C138.293 48.1451 137.565 48.2544 138.21 47.6133C139.642 46.1873 140.448 44.3024 140.471 42.3521L140.486 41.5428L140.829 41.529C141.215 41.5136 141.65 41.3095 141.81 41.0611C142.078 40.6633 141.767 40.6067 141.408 40.9874C141.211 41.1957 141.121 41.2408 140.862 41.2512C140.609 41.2613 140.51 41.2238 140.335 41.058C140.132 40.8656 140.124 40.8314 140.102 40.2032C140.08 39.5543 140.079 39.5474 139.912 39.5749C139.779 39.594 139.726 39.6791 139.639 40.0006C139.526 40.4129 139.315 40.7878 139.083 40.9906C138.708 41.3167 138.22 41.0113 138.121 40.3862C138.085 40.1802 138.074 39.842 138.088 39.6341C138.102 39.4261 138.102 39.2326 138.087 39.1986C138.035 39.1177 137.783 39.1554 137.758 39.2463C137.746 39.2952 137.685 39.5534 137.63 39.8183C137.496 40.4112 137.153 40.8466 136.803 40.8606C136.48 40.8736 136.338 40.6373 136.266 39.9558C136.221 39.4737 136.19 39.3989 136.085 39.4031C135.994 39.4068 135.92 39.5134 135.832 39.7935C135.528 40.7527 134.758 41.2466 134.265 40.7963C134.084 40.6308 134.075 40.5827 134.045 39.9134C134.024 39.3266 133.999 39.2032 133.907 39.1931C133.844 39.1818 133.741 39.2619 133.668 39.3686C133.181 40.1208 132.598 40.5381 132.132 40.4669C131.532 40.3734 131.078 39.8248 131.143 39.2691C131.265 38.3241 132.434 37.2128 133.768 36.793C134.393 36.5952 136.722 36.4951 137.675 36.6367C138.938 36.8142 139.847 37.1096 140.722 37.6345C141.863 38.3147 142.474 39.1198 142.716 40.2507C143.162 42.355 141.964 45.7072 140.059 47.6707C139.699 48.0376 139.675 48.1423 139.941 48.1316C140.201 48.1212 141.105 47.0619 141.696 46.0843C143.218 43.5694 143.62 40.5809 142.651 38.9744C141.854 37.6446 140.058 36.6934 137.757 36.3845C136.868 36.268 134.825 36.2737 134.197 36.3957Z" fill="#7E36E3" stroke="#7E36E3" strokeWidth="0.8"/>
-            <path d="M149.141 39.3036C148.327 39.4681 147.66 39.8217 147.114 40.392C146.687 40.8363 146.495 41.2183 146.509 41.6186C146.522 41.9819 146.654 42.2297 146.925 42.4088C147.393 42.7117 148.001 42.6715 148.453 42.3107L148.684 42.1221L148.723 42.3368C148.775 42.5878 148.957 42.8495 149.143 42.9317C149.483 43.0816 150.075 42.9102 150.339 42.578C150.468 42.4199 150.468 42.4199 150.531 42.5228C150.583 42.6104 150.575 42.7056 150.492 42.962C150.032 44.4728 148.059 46.0381 145.171 47.2136C144.57 47.4539 143.031 48.2906 143.344 48.2781C143.57 48.269 146.132 47.0758 147.006 46.6137C148.907 45.6095 150.422 44.2041 150.736 43.1579L150.819 42.8962L151.137 42.8888C151.325 42.8812 151.481 42.8539 151.523 42.8101C151.725 42.5963 151.717 42.8128 151.504 43.5227C150.977 45.2735 149.901 46.7562 148.401 47.8182C148.006 48.0977 147.889 48.0614 147.993 48.0994C148.08 48.1328 148.257 48.172 148.51 48.0089C149.634 47.2837 150.769 45.9621 151.367 44.6989C151.573 44.2582 151.953 43.104 151.948 42.9354C151.943 42.8091 152.008 42.8065 152.136 42.9332C152.393 43.1918 152.915 43.1446 153.217 42.8319L153.389 42.6563L153.47 42.8376C153.517 42.9412 153.585 43.0386 153.618 43.0531C153.695 43.0975 153.717 43.7189 153.664 44.2167C153.482 45.8324 152.655 47.3157 151.409 48.2568C151.39 48.2674 152.137 48.2192 152.284 48.2661C152.284 48.2661 151.725 48.3496 152.221 47.8604C153.321 46.7722 153.94 45.3342 153.958 43.8464L153.969 43.2289L154.233 43.2184C154.529 43.2065 154.864 43.0508 154.987 42.8613C155.192 42.5577 154.953 42.5146 154.677 42.8051C154.526 42.9641 154.457 42.9985 154.258 43.0064C154.064 43.0142 153.988 42.9856 153.854 42.8591C153.698 42.7125 153.692 42.6863 153.675 42.2071C153.658 41.7121 153.658 41.7068 153.529 41.7278C153.427 41.7424 153.386 41.8074 153.319 42.0526C153.233 42.3672 153.07 42.6532 152.892 42.808C152.604 43.0568 152.229 42.824 152.154 42.3471C152.127 42.19 152.118 41.932 152.128 41.7733C152.139 41.6147 152.139 41.467 152.128 41.4411C152.088 41.3795 151.895 41.4083 151.875 41.4776C151.866 41.5149 151.819 41.7119 151.777 41.9139C151.674 42.3663 151.411 42.6985 151.142 42.7093C150.894 42.7192 150.785 42.539 150.729 42.0192C150.695 41.6514 150.671 41.5943 150.591 41.5976C150.521 41.6004 150.464 41.6817 150.396 41.8954C150.163 42.6272 149.572 43.0042 149.193 42.6607C149.054 42.5345 149.047 42.4978 149.024 41.9872C149.009 41.5396 148.989 41.4455 148.919 41.4377C148.87 41.4291 148.791 41.4903 148.735 41.5717C148.361 42.1456 147.913 42.4641 147.556 42.4099C147.095 42.3387 146.746 41.9202 146.796 41.4963C146.89 40.7754 147.788 39.9274 148.812 39.6069C149.292 39.4558 151.08 39.379 151.812 39.4868C152.782 39.622 153.48 39.8472 154.152 40.2474C155.027 40.7661 155.496 41.3801 155.682 42.2428C156.024 43.8481 155.104 46.4056 153.64 47.9038C153.364 48.1838 153.346 48.2636 153.55 48.2554C153.749 48.2475 154.444 47.4392 154.898 46.6933C156.067 44.7745 156.376 42.4945 155.633 41.2692C155.02 40.2549 153.642 39.5296 151.875 39.2944C151.193 39.2057 149.624 39.2105 149.141 39.3036Z" fill="#7E36E3" stroke="#7E36E3" strokeWidth="0.8"/>
-            <path d="M148.329 1.10201C148.248 1.2363 147.724 3.46551 147.711 3.70723C147.711 3.77438 147.872 3.80123 148.074 3.77438C148.409 3.73409 148.45 3.65351 148.678 2.65977C149.027 1.10201 149.027 1.06172 148.691 1.008C148.544 0.981145 148.369 1.02143 148.329 1.10201Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M141.667 1.85453C141.667 1.90824 141.949 2.41855 142.298 3.00942C142.795 3.81516 143.01 4.07031 143.238 4.07031C143.399 4.07031 143.56 4.03003 143.587 3.96288C143.628 3.90916 143.372 3.41229 143.023 2.8617C142.526 2.05596 142.325 1.8411 142.03 1.81424C141.828 1.78738 141.667 1.81424 141.667 1.85453Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M152.706 4.74171C152.089 5.17143 151.592 5.5743 151.592 5.64145C151.592 5.69517 151.74 5.76231 151.901 5.78917C152.223 5.81603 154.291 4.41941 154.398 4.09712C154.519 3.70767 153.754 4.00311 152.706 4.74171Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M145.575 6.97098C144.353 7.29328 143.52 8.08559 142.997 9.40163C142.741 10.0462 142.688 10.4357 142.688 11.7248C142.674 13.0946 142.715 13.3498 143.01 13.9406C143.789 15.5253 145.28 16.4519 147.106 16.4519C150.074 16.4787 152.101 14.6792 151.819 12.2889C151.564 10.0865 149.456 8.73018 147.294 9.34791C146.072 9.7105 145.387 10.6237 145.387 11.886C145.387 12.4903 145.454 12.7589 145.736 13.1886C146.528 14.3838 148.019 14.5718 148.852 13.5915C149.402 12.9335 149.241 11.8591 148.597 11.8591C148.153 11.8591 148.113 11.9666 148.355 12.3963C148.529 12.7052 148.543 12.8395 148.409 13.1215C148.261 13.4438 148.18 13.4706 147.455 13.4706C146.743 13.4706 146.649 13.4303 146.421 13.108C146.112 12.638 146.018 11.5771 146.233 11.0534C146.609 10.1537 147.912 9.45535 148.825 9.65678C149.483 9.79107 150.221 10.3282 150.624 10.9325C151.148 11.7248 151.229 13.0677 150.799 13.9272C150.168 15.2432 148.771 16.0221 147.052 16.0221C145.481 16.0221 144.42 15.3104 143.735 13.8063C143.131 12.4634 143.211 10.664 143.937 9.09276C144.595 7.66929 145.481 7.23956 147.549 7.30671C149.026 7.36042 149.389 7.17242 148.247 6.93069C147.24 6.72926 146.434 6.74269 145.575 6.97098Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M137.236 7.88366C137.236 8.1791 137.317 8.19253 139.09 8.29996C140.365 8.36711 140.594 8.35368 140.594 8.1791C140.594 7.88366 140.526 7.87023 138.821 7.7628C137.411 7.68223 137.236 7.69566 137.236 7.88366Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M156.131 8.24625C155.997 8.32683 155.433 8.47455 154.869 8.56855C153.808 8.75655 153.633 8.83713 153.835 9.03856C154.117 9.32057 156.977 8.64912 156.977 8.28654C156.977 8.05825 156.48 8.04482 156.131 8.24625Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M154.56 14.3434C154.439 14.4911 154.479 14.5717 154.801 14.7328C155.58 15.1357 157.259 15.7534 157.581 15.7534C157.984 15.7534 158.159 15.5117 157.823 15.4043C157.528 15.3103 155.862 14.6522 155.204 14.3568C154.788 14.1822 154.694 14.1822 154.56 14.3434Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M139.586 15.0686C139.291 15.2029 138.633 15.4177 138.109 15.552C137.075 15.8206 136.806 15.9683 137.129 16.0892C137.558 16.2637 140.862 15.2297 140.862 14.9343C140.862 14.7328 140.163 14.8134 139.586 15.0686Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M151.538 18.0364C151.431 18.1975 151.74 18.4661 152.975 19.339C153.741 19.8627 153.929 19.9433 154.184 19.8224C154.466 19.7015 154.439 19.6747 153.808 19.245C153.445 18.9898 152.908 18.5869 152.599 18.3452C152.035 17.8886 151.686 17.7946 151.538 18.0364Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M142.755 20.8024C142.406 21.9573 142.352 22.2662 142.5 22.3468C142.742 22.5079 142.822 22.4945 142.97 22.293C143.158 22.0647 143.937 19.5401 143.843 19.4461C143.803 19.4058 143.641 19.3789 143.48 19.3789C143.225 19.3789 143.131 19.567 142.755 20.8024Z" fill="#1F2937" stroke="#7E36E3"/>
-            <path d="M147.898 19.8492C147.791 20.1044 148.046 21.3667 148.328 22.0381C148.476 22.3873 148.583 22.4813 148.811 22.441C149.16 22.3873 149.16 22.4007 148.758 20.8564C148.529 20.0372 148.382 19.7015 148.207 19.6746C148.086 19.6478 147.938 19.7284 147.898 19.8492Z" fill="#1F2937" stroke="#7E36E3"/>
-            </svg>
-          </div>
-        </div>
-      </div>
-
-      {/* Profile Content */}
-      <div className="bg-white p-4 flex items-center h-full">
-      <div className="bg-gray-100 rounded-full p-0.5 mr-4">
-          {/* <ProfileEditIcon className="w-12 h-12" /> */}
-          <svg width="56" height="56" viewBox="0 0 56 56" fill="none" xmlns="http://www.w3.org/2000/svg">
-          <g clipPath="url(#clip0_2969_45783)">
-          <circle cx="28" cy="28" r="27" fill="white" stroke="#E5E7EB" strokeWidth={2}/>
-          <path fillRule="evenodd" clipRule="evenodd" d="M44.047 34.9929C41.3481 41.1773 35.1789 45.5 28 45.5C20.3796 45.5 13.8967 40.6292 11.495 33.8311C8.35263 33.7404 5.83311 31.1645 5.83311 28.0002C5.83311 24.836 8.35254 22.2602 11.4948 22.1694C13.8964 15.371 20.3794 10.5 28 10.5C36.0579 10.5 42.8438 15.9461 44.8777 23.3582C47.8441 23.6327 50.1667 26.1283 50.1667 29.1664C50.1667 32.3881 47.555 34.9998 44.3333 34.9998C44.2373 34.9998 44.1418 34.9974 44.047 34.9929Z" fill="#E5E7EB"/>
-          <path fillRule="evenodd" clipRule="evenodd" d="M26.3669 25.2016C25.8157 24.0949 24.6192 23.333 23.2081 23.333C22.5902 23.333 21.6353 23.7465 21.2421 24.4184C21.186 24.5735 20.9613 24.5735 20.9051 24.4184C20.4558 23.7982 19.557 23.3847 18.883 23.3847C17.0293 23.3847 15.4566 24.7802 15.4004 26.4859V26.5893C15.4004 27.4679 15.7936 28.2949 16.4676 28.9152V28.9669C16.5238 29.0185 19.22 31.1894 20.4558 32.1714C20.7928 32.4299 21.2983 32.4299 21.6353 32.1714C22.1544 31.7589 22.9213 31.1367 23.6445 30.5498L23.6445 30.5498C24.643 29.7395 25.5583 28.9968 25.6235 28.9669V28.9152C26.2195 28.4124 26.596 27.7076 26.6751 26.9078C27.0642 26.8601 27.5041 26.8331 28.0007 26.8331C28.7553 26.8331 29.3792 26.8954 29.8925 26.9983C29.8916 27.0155 29.8909 27.0327 29.8903 27.05V27.1533C29.8903 28.032 30.2835 28.859 30.9575 29.4792V29.5309C31.0137 29.5826 33.7099 31.7535 34.9457 32.7355C35.2827 32.9939 35.7882 32.9939 36.1252 32.7355C36.6443 32.323 37.4112 31.7007 38.1344 31.1139C39.1329 30.3036 40.0482 29.5609 40.1134 29.5309V29.4792C40.7874 28.9107 41.1806 28.0837 41.1806 27.1533V27.05C41.1244 25.2926 39.6078 23.8971 37.698 23.8971C37.0801 23.8971 36.1252 24.3106 35.732 24.9825C35.6759 25.1376 35.4512 25.1376 35.395 24.9825C34.9457 24.3623 34.0469 23.9488 33.3729 23.9488C32.1794 23.9488 31.1023 24.5273 30.4716 25.3946C29.8707 25.2072 29.066 25.0831 28.0007 25.0831C27.3682 25.0831 26.8276 25.1268 26.3669 25.2016Z" fill="#6B7280"/>
-          <path d="M25.083 33.25C25.2775 33.6389 26.133 34.4167 27.9997 34.4167C29.8663 34.4167 30.7219 33.6389 30.9163 33.25" stroke="#6B7280" strokeWidth={1.75} strokeLinecap="round" strokeLinejoin="round"/>
-          <path d="M25.1911 13.0162C24.5791 13.4958 23.6168 15.4282 24.147 16.7875C24.6771 18.1469 25.5062 18.4837 25.9272 18.5919" stroke="#6B7280" strokeWidth={1.39112} strokeLinecap="round"/>
-          <path d="M28.9101 13.9739C28.2972 14.3917 27.857 16.1891 28.4439 17.445C28.8807 18.3796 29.9979 18.7424 30.3536 18.8338" stroke="#6B7280" strokeWidth={1.39112} strokeLinecap="round"/>
-          <path d="M32.0018 14.6125C31.7083 14.9974 31.9234 16.1068 32.4277 16.8577C32.7258 17.3017 33.6286 17.9501 33.8676 17.9335" stroke="#6B7280" strokeWidth={1.38833} strokeLinecap="round"/>
-          <circle cx="45" cy="45" r="10" fill="#E5E7EB" stroke="white" strokeWidth={2}/>
-          <path d="M45.0118 41.1895C45.1164 40.964 45.3836 40.8668 45.6086 40.9723L47.5454 41.8803C47.7705 41.9858 47.8681 42.2541 47.7635 42.4795L45.0823 48.2603C45.0314 48.3701 44.9385 48.4546 44.8245 48.4947L43.4156 48.9904C43.1852 49.0714 42.932 48.9527 42.8456 48.7231L42.3177 47.3195C42.275 47.2059 42.2796 47.0801 42.3306 46.9703L45.0118 41.1895Z" fill="#9CA3AF"/>
-          <path d="M43.126 41.3809L48.4896 43.8973" stroke="#E5E7EB" strokeWidth="0.5625"/>
-          </g>
-          <defs>
-          <clipPath id="clip0_2969_45783">
-          <rect width="56" height="56" fill="white"/>
-          </clipPath>
-          </defs>
-          </svg>
-        </div>
-        <div>
-          <div className="text-gray-800 font-bold text-md">홍길동</div>
-          <div className="text-sm text-gray-800 flex gap-2">
-            <div className="font-bold">company.</div>
-            <div>코드잇</div>
-          </div>
-          <div className="text-sm text-gray-800 flex gap-2">
-            <div className="font-bold">E-mail.</div>
-            <div> codeit@codeit.com</div>
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
