+++ src/app/mypage/my-meetings/page.tsx
@@ -1,7 +1,6 @@
 "use client";
 import { useState } from "react";
-import ProfileCard from "../components/ProfileCard"
-import Navbar from "@/components/shared/ui/Navbar";
+import ProfileCard from "../../../components/mypage/ProfileCard"
 import Image from "next/image";
 
 
@@ -9,7 +8,6 @@ export default function MyPage() {
     const [selectedTab, setSelectedTab] = useState(0);
     return (
         <div className="bg-gray-50 min-h-screen">
-            <Navbar />
             {/* My Page Section */}
             <div className="bg-gray-50 min-h-screen">
                 <div className="bg-gray-30 min-h-screen">
@@ -50,20 +48,12 @@ export default function MyPage() {
                                 <div>
                                     <div className="flex gap-2 py-2">
                                         <div>
-                                            {/*이용예정.svg*/}
-                                            <svg width="76" height="32" viewBox="0 0 76 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <rect width="76" height="32" rx="16" fill="#D3C1FF"/>
-                                            <path d="M22.5957 9.91211V22.1895H21.2969V9.91211H22.5957ZM16.0469 10.8281C17.8652 10.8281 19.2051 12.4961 19.2051 15.1211C19.2051 17.7598 17.8652 19.4277 16.0469 19.4277C14.2012 19.4277 12.875 17.7598 12.875 15.1211C12.875 12.4961 14.2012 10.8281 16.0469 10.8281ZM16.0469 11.9902C14.9121 11.9902 14.1328 13.1934 14.1328 15.1211C14.1328 17.0625 14.9121 18.2656 16.0469 18.2656C17.168 18.2656 17.9473 17.0625 17.9473 15.1211C17.9473 13.1934 17.168 11.9902 16.0469 11.9902ZM35.7207 15.8867V16.9395H24.5508V15.8867H27.2305V14.2324C26.2871 13.8496 25.7539 13.2344 25.7539 12.4277C25.7539 10.9785 27.4355 10.1309 30.1289 10.1309C32.8086 10.1309 34.5039 10.9785 34.5039 12.4277C34.5039 13.2344 33.9639 13.8564 33.0137 14.2324V15.8867H35.7207ZM30.1289 17.7734C32.7676 17.7734 34.3535 18.5527 34.3535 19.9473C34.3535 21.3691 32.7676 22.1348 30.1289 22.1348C27.4629 22.1348 25.877 21.3691 25.877 19.9473C25.877 18.5527 27.4629 17.7734 30.1289 17.7734ZM30.1289 18.7715C28.2559 18.7715 27.1895 19.1953 27.1895 19.9473C27.1895 20.7129 28.2559 21.1504 30.1289 21.1367C31.9746 21.1504 33.0547 20.7129 33.0547 19.9473C33.0547 19.1953 31.9746 18.7715 30.1289 18.7715ZM30.1289 11.1289C28.2285 11.1289 27.0801 11.6074 27.0938 12.4277C27.0801 13.248 28.2285 13.7129 30.1289 13.7129C32.0293 13.7129 33.1641 13.248 33.1641 12.4277C33.1641 11.6074 32.0293 11.1289 30.1289 11.1289ZM28.543 15.8867H31.7285V14.5742C31.25 14.6562 30.71 14.6973 30.1289 14.6973C29.5547 14.6973 29.0215 14.6562 28.543 14.5879V15.8867ZM50.6094 9.91211V22.1621H49.3652V9.91211H50.6094ZM48.1484 10.1445V21.5605H46.918V17.6094H45.1406C44.7031 18.8262 43.8896 19.5303 42.8438 19.5234C41.2441 19.5371 40.2051 17.8555 40.2051 15.2031C40.2051 12.5508 41.2441 10.8828 42.8438 10.8965C43.8896 10.8896 44.6963 11.5869 45.1406 12.7969H46.918V10.1445H48.1484ZM42.8438 12.1133C41.9414 12.1133 41.3945 13.248 41.3945 15.2031C41.3945 17.1582 41.9414 18.293 42.8438 18.293C43.7461 18.293 44.3203 17.1582 44.3203 15.2031C44.3203 13.248 43.7461 12.1133 42.8438 12.1133ZM45.4141 13.8496C45.4756 14.2666 45.5098 14.7178 45.5098 15.2031C45.5098 15.6816 45.4756 16.1328 45.4141 16.543H46.918V13.8496H45.4141ZM62.3809 9.91211V17.1992H61.082V14.0957H58.7988V13.0293H61.082V9.91211H62.3809ZM58.2793 17.541C60.8496 17.541 62.4219 18.3887 62.4219 19.8516C62.4219 21.3145 60.8496 22.1484 58.2793 22.1621C55.709 22.1484 54.123 21.3145 54.1367 19.8516C54.123 18.3887 55.709 17.541 58.2793 17.541ZM58.2793 18.5527C56.4883 18.5527 55.4219 19.0176 55.4219 19.8516C55.4219 20.6719 56.4883 21.1504 58.2793 21.1367C60.0703 21.1504 61.1367 20.6719 61.1367 19.8516C61.1367 19.0176 60.0703 18.5527 58.2793 18.5527ZM56.5566 12.0723C56.5566 13.6309 57.6641 15.1348 59.5098 15.7637L58.8398 16.7891C57.4658 16.3105 56.4541 15.333 55.9277 14.123C55.3877 15.4971 54.3145 16.5977 52.8516 17.1309L52.168 16.0918C54.0684 15.4355 55.2305 13.7812 55.2305 12.0859V11.8125H52.5781V10.7598H59.1816V11.8125H56.5566V12.0723Z" fill="#7E36E3"/>
-                                            </svg>
+                                            {/*scheduled.svg*/}
+                                            <img src="/icons/scheduled.svg" alt="이용예정" width={72} height={120} />
                                         </div>
                                         <div>
-                                            {/*개설확정.svg*/}
-                                            <svg width="93" height="32" viewBox="0 0 93 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <rect x="0.5" y="0.5" width="92" height="31" rx="15.5" fill="white"/>
-                                            <rect x="0.5" y="0.5" width="92" height="31" rx="15.5" stroke="#D3C1FF"/>
-                                            <path d="M16 15.2456L19.0877 18.3333L24.4209 13" stroke="#A065FF" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"/>
-                                            <path d="M42.9375 9.91211V22.1621H41.6934V15.8047H40.1758V21.5742H38.959V10.2402H40.1758V14.752H41.6934V9.91211H42.9375ZM37.7422 11.4707C37.7285 14.6973 36.6621 17.582 33.1621 19.5645L32.4102 18.6621C35.1309 17.124 36.2793 15.0391 36.4844 12.5371H32.9297V11.4707H37.7422ZM54.7637 9.91211V16.2148H53.4648V13.0703H50.8809V12.0176H53.4648V9.91211H54.7637ZM54.7637 16.7891V19.8379H48.0508V20.9727H55.1875V22.0117H46.7793V18.8535H53.4785V17.8145H46.752V16.7891H54.7637ZM48.8574 11.375C48.8574 12.9336 49.9785 14.3828 51.8242 14.9707L51.1406 15.9824C49.7939 15.5176 48.7754 14.5947 48.2422 13.4258C47.7021 14.7383 46.6494 15.7705 45.207 16.2695L44.5098 15.2305C46.4238 14.6152 47.5586 13.0977 47.5723 11.375V10.3086H48.8574V11.375ZM66.2891 9.92578V13.5625H68.0391V14.6289H66.2891V18.1699H64.9902V9.92578H66.2891ZM56.5684 16.8027C57.5186 16.8027 58.6191 16.8027 59.7676 16.7617V15.9414C58.373 15.7979 57.4844 15.1416 57.498 14.1504C57.4844 13.0293 58.6465 12.332 60.4102 12.3457C62.1602 12.332 63.3223 13.0293 63.3223 14.1504C63.3223 15.1348 62.4473 15.791 61.0664 15.9414V16.7207C62.1602 16.666 63.2676 16.5771 64.293 16.4336L64.4023 17.3633C61.7637 17.8281 58.8652 17.8418 56.7051 17.8418L56.5684 16.8027ZM66.2891 18.7441V22.1621H64.9902V19.7559H58.0996V18.7441H66.2891ZM63.9512 10.9648V11.9355H56.8828V10.9648H59.7676V9.81641H61.0664V10.9648H63.9512ZM60.4102 13.2344C59.3574 13.2207 58.7285 13.5625 58.7285 14.1504C58.7285 14.7246 59.3574 15.0938 60.4102 15.0938C61.4492 15.0938 62.0781 14.7246 62.0781 14.1504C62.0781 13.5625 61.4492 13.2207 60.4102 13.2344ZM78.9629 9.91211V17.1992H77.6641V14.0957H75.3809V13.0293H77.6641V9.91211H78.9629ZM74.8613 17.541C77.4316 17.541 79.0039 18.3887 79.0039 19.8516C79.0039 21.3145 77.4316 22.1484 74.8613 22.1621C72.291 22.1484 70.7051 21.3145 70.7188 19.8516C70.7051 18.3887 72.291 17.541 74.8613 17.541ZM74.8613 18.5527C73.0703 18.5527 72.0039 19.0176 72.0039 19.8516C72.0039 20.6719 73.0703 21.1504 74.8613 21.1367C76.6523 21.1504 77.7188 20.6719 77.7188 19.8516C77.7188 19.0176 76.6523 18.5527 74.8613 18.5527ZM73.1387 12.0723C73.1387 13.6309 74.2461 15.1348 76.0918 15.7637L75.4219 16.7891C74.0479 16.3105 73.0361 15.333 72.5098 14.123C71.9697 15.4971 70.8965 16.5977 69.4336 17.1309L68.75 16.0918C70.6504 15.4355 71.8125 13.7812 71.8125 12.0859V11.8125H69.1602V10.7598H75.7637V11.8125H73.1387V12.0723Z" fill="#A065FF"/>
-                                            </svg>
+                                            {/*sonfirmed.svg*/}
+                                            <img src="/icons/confirmed.svg" alt="개설확정" width={90} height={120} />
                                         </div>
                                     </div>
                                         <h3 className="py-2">달램핏 오피스 스트레칭 | <span>을지로 3가</span> </h3>
@@ -90,20 +80,12 @@ export default function MyPage() {
                                 <div>
                                     <div className="flex gap-2 py-2">
                                         <div>
-                                            {/*이용예정.svg*/}
-                                            <svg width="76" height="32" viewBox="0 0 76 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <rect width="76" height="32" rx="16" fill="#D3C1FF"/>
-                                            <path d="M22.5957 9.91211V22.1895H21.2969V9.91211H22.5957ZM16.0469 10.8281C17.8652 10.8281 19.2051 12.4961 19.2051 15.1211C19.2051 17.7598 17.8652 19.4277 16.0469 19.4277C14.2012 19.4277 12.875 17.7598 12.875 15.1211C12.875 12.4961 14.2012 10.8281 16.0469 10.8281ZM16.0469 11.9902C14.9121 11.9902 14.1328 13.1934 14.1328 15.1211C14.1328 17.0625 14.9121 18.2656 16.0469 18.2656C17.168 18.2656 17.9473 17.0625 17.9473 15.1211C17.9473 13.1934 17.168 11.9902 16.0469 11.9902ZM35.7207 15.8867V16.9395H24.5508V15.8867H27.2305V14.2324C26.2871 13.8496 25.7539 13.2344 25.7539 12.4277C25.7539 10.9785 27.4355 10.1309 30.1289 10.1309C32.8086 10.1309 34.5039 10.9785 34.5039 12.4277C34.5039 13.2344 33.9639 13.8564 33.0137 14.2324V15.8867H35.7207ZM30.1289 17.7734C32.7676 17.7734 34.3535 18.5527 34.3535 19.9473C34.3535 21.3691 32.7676 22.1348 30.1289 22.1348C27.4629 22.1348 25.877 21.3691 25.877 19.9473C25.877 18.5527 27.4629 17.7734 30.1289 17.7734ZM30.1289 18.7715C28.2559 18.7715 27.1895 19.1953 27.1895 19.9473C27.1895 20.7129 28.2559 21.1504 30.1289 21.1367C31.9746 21.1504 33.0547 20.7129 33.0547 19.9473C33.0547 19.1953 31.9746 18.7715 30.1289 18.7715ZM30.1289 11.1289C28.2285 11.1289 27.0801 11.6074 27.0938 12.4277C27.0801 13.248 28.2285 13.7129 30.1289 13.7129C32.0293 13.7129 33.1641 13.248 33.1641 12.4277C33.1641 11.6074 32.0293 11.1289 30.1289 11.1289ZM28.543 15.8867H31.7285V14.5742C31.25 14.6562 30.71 14.6973 30.1289 14.6973C29.5547 14.6973 29.0215 14.6562 28.543 14.5879V15.8867ZM50.6094 9.91211V22.1621H49.3652V9.91211H50.6094ZM48.1484 10.1445V21.5605H46.918V17.6094H45.1406C44.7031 18.8262 43.8896 19.5303 42.8438 19.5234C41.2441 19.5371 40.2051 17.8555 40.2051 15.2031C40.2051 12.5508 41.2441 10.8828 42.8438 10.8965C43.8896 10.8896 44.6963 11.5869 45.1406 12.7969H46.918V10.1445H48.1484ZM42.8438 12.1133C41.9414 12.1133 41.3945 13.248 41.3945 15.2031C41.3945 17.1582 41.9414 18.293 42.8438 18.293C43.7461 18.293 44.3203 17.1582 44.3203 15.2031C44.3203 13.248 43.7461 12.1133 42.8438 12.1133ZM45.4141 13.8496C45.4756 14.2666 45.5098 14.7178 45.5098 15.2031C45.5098 15.6816 45.4756 16.1328 45.4141 16.543H46.918V13.8496H45.4141ZM62.3809 9.91211V17.1992H61.082V14.0957H58.7988V13.0293H61.082V9.91211H62.3809ZM58.2793 17.541C60.8496 17.541 62.4219 18.3887 62.4219 19.8516C62.4219 21.3145 60.8496 22.1484 58.2793 22.1621C55.709 22.1484 54.123 21.3145 54.1367 19.8516C54.123 18.3887 55.709 17.541 58.2793 17.541ZM58.2793 18.5527C56.4883 18.5527 55.4219 19.0176 55.4219 19.8516C55.4219 20.6719 56.4883 21.1504 58.2793 21.1367C60.0703 21.1504 61.1367 20.6719 61.1367 19.8516C61.1367 19.0176 60.0703 18.5527 58.2793 18.5527ZM56.5566 12.0723C56.5566 13.6309 57.6641 15.1348 59.5098 15.7637L58.8398 16.7891C57.4658 16.3105 56.4541 15.333 55.9277 14.123C55.3877 15.4971 54.3145 16.5977 52.8516 17.1309L52.168 16.0918C54.0684 15.4355 55.2305 13.7812 55.2305 12.0859V11.8125H52.5781V10.7598H59.1816V11.8125H56.5566V12.0723Z" fill="#7E36E3"/>
-                                            </svg>
+                                            {/*scheduled.svg*/}
+                                            <img src="/icons/scheduled.svg" alt="이용예정" width={72} height={120} />
                                         </div>
                                         <div>
-                                            {/*개설확정.svg*/}
-                                            <svg width="93" height="32" viewBox="0 0 93 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <rect x="0.5" y="0.5" width="92" height="31" rx="15.5" fill="white"/>
-                                            <rect x="0.5" y="0.5" width="92" height="31" rx="15.5" stroke="#D3C1FF"/>
-                                            <path d="M16 15.2456L19.0877 18.3333L24.4209 13" stroke="#A065FF" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"/>
-                                            <path d="M42.9375 9.91211V22.1621H41.6934V15.8047H40.1758V21.5742H38.959V10.2402H40.1758V14.752H41.6934V9.91211H42.9375ZM37.7422 11.4707C37.7285 14.6973 36.6621 17.582 33.1621 19.5645L32.4102 18.6621C35.1309 17.124 36.2793 15.0391 36.4844 12.5371H32.9297V11.4707H37.7422ZM54.7637 9.91211V16.2148H53.4648V13.0703H50.8809V12.0176H53.4648V9.91211H54.7637ZM54.7637 16.7891V19.8379H48.0508V20.9727H55.1875V22.0117H46.7793V18.8535H53.4785V17.8145H46.752V16.7891H54.7637ZM48.8574 11.375C48.8574 12.9336 49.9785 14.3828 51.8242 14.9707L51.1406 15.9824C49.7939 15.5176 48.7754 14.5947 48.2422 13.4258C47.7021 14.7383 46.6494 15.7705 45.207 16.2695L44.5098 15.2305C46.4238 14.6152 47.5586 13.0977 47.5723 11.375V10.3086H48.8574V11.375ZM66.2891 9.92578V13.5625H68.0391V14.6289H66.2891V18.1699H64.9902V9.92578H66.2891ZM56.5684 16.8027C57.5186 16.8027 58.6191 16.8027 59.7676 16.7617V15.9414C58.373 15.7979 57.4844 15.1416 57.498 14.1504C57.4844 13.0293 58.6465 12.332 60.4102 12.3457C62.1602 12.332 63.3223 13.0293 63.3223 14.1504C63.3223 15.1348 62.4473 15.791 61.0664 15.9414V16.7207C62.1602 16.666 63.2676 16.5771 64.293 16.4336L64.4023 17.3633C61.7637 17.8281 58.8652 17.8418 56.7051 17.8418L56.5684 16.8027ZM66.2891 18.7441V22.1621H64.9902V19.7559H58.0996V18.7441H66.2891ZM63.9512 10.9648V11.9355H56.8828V10.9648H59.7676V9.81641H61.0664V10.9648H63.9512ZM60.4102 13.2344C59.3574 13.2207 58.7285 13.5625 58.7285 14.1504C58.7285 14.7246 59.3574 15.0938 60.4102 15.0938C61.4492 15.0938 62.0781 14.7246 62.0781 14.1504C62.0781 13.5625 61.4492 13.2207 60.4102 13.2344ZM78.9629 9.91211V17.1992H77.6641V14.0957H75.3809V13.0293H77.6641V9.91211H78.9629ZM74.8613 17.541C77.4316 17.541 79.0039 18.3887 79.0039 19.8516C79.0039 21.3145 77.4316 22.1484 74.8613 22.1621C72.291 22.1484 70.7051 21.3145 70.7188 19.8516C70.7051 18.3887 72.291 17.541 74.8613 17.541ZM74.8613 18.5527C73.0703 18.5527 72.0039 19.0176 72.0039 19.8516C72.0039 20.6719 73.0703 21.1504 74.8613 21.1367C76.6523 21.1504 77.7188 20.6719 77.7188 19.8516C77.7188 19.0176 76.6523 18.5527 74.8613 18.5527ZM73.1387 12.0723C73.1387 13.6309 74.2461 15.1348 76.0918 15.7637L75.4219 16.7891C74.0479 16.3105 73.0361 15.333 72.5098 14.123C71.9697 15.4971 70.8965 16.5977 69.4336 17.1309L68.75 16.0918C70.6504 15.4355 71.8125 13.7812 71.8125 12.0859V11.8125H69.1602V10.7598H75.7637V11.8125H73.1387V12.0723Z" fill="#A065FF"/>
-                                            </svg>
+                                            {/*sonfirmed.svg*/}
+                                            <img src="/icons/confirmed.svg" alt="개설확정" width={90} height={120} />
                                         </div>
                                     </div>
                                         <h3 className="py-2">달램핏 오피스 스트레칭 | <span>을지로 3가</span> </h3>
@@ -130,20 +112,12 @@ export default function MyPage() {
                                 <div>
                                     <div className="flex gap-2 py-2">
                                         <div>
-                                            {/*이용예정.svg*/}
-                                            <svg width="76" height="32" viewBox="0 0 76 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <rect width="76" height="32" rx="16" fill="#D3C1FF"/>
-                                            <path d="M22.5957 9.91211V22.1895H21.2969V9.91211H22.5957ZM16.0469 10.8281C17.8652 10.8281 19.2051 12.4961 19.2051 15.1211C19.2051 17.7598 17.8652 19.4277 16.0469 19.4277C14.2012 19.4277 12.875 17.7598 12.875 15.1211C12.875 12.4961 14.2012 10.8281 16.0469 10.8281ZM16.0469 11.9902C14.9121 11.9902 14.1328 13.1934 14.1328 15.1211C14.1328 17.0625 14.9121 18.2656 16.0469 18.2656C17.168 18.2656 17.9473 17.0625 17.9473 15.1211C17.9473 13.1934 17.168 11.9902 16.0469 11.9902ZM35.7207 15.8867V16.9395H24.5508V15.8867H27.2305V14.2324C26.2871 13.8496 25.7539 13.2344 25.7539 12.4277C25.7539 10.9785 27.4355 10.1309 30.1289 10.1309C32.8086 10.1309 34.5039 10.9785 34.5039 12.4277C34.5039 13.2344 33.9639 13.8564 33.0137 14.2324V15.8867H35.7207ZM30.1289 17.7734C32.7676 17.7734 34.3535 18.5527 34.3535 19.9473C34.3535 21.3691 32.7676 22.1348 30.1289 22.1348C27.4629 22.1348 25.877 21.3691 25.877 19.9473C25.877 18.5527 27.4629 17.7734 30.1289 17.7734ZM30.1289 18.7715C28.2559 18.7715 27.1895 19.1953 27.1895 19.9473C27.1895 20.7129 28.2559 21.1504 30.1289 21.1367C31.9746 21.1504 33.0547 20.7129 33.0547 19.9473C33.0547 19.1953 31.9746 18.7715 30.1289 18.7715ZM30.1289 11.1289C28.2285 11.1289 27.0801 11.6074 27.0938 12.4277C27.0801 13.248 28.2285 13.7129 30.1289 13.7129C32.0293 13.7129 33.1641 13.248 33.1641 12.4277C33.1641 11.6074 32.0293 11.1289 30.1289 11.1289ZM28.543 15.8867H31.7285V14.5742C31.25 14.6562 30.71 14.6973 30.1289 14.6973C29.5547 14.6973 29.0215 14.6562 28.543 14.5879V15.8867ZM50.6094 9.91211V22.1621H49.3652V9.91211H50.6094ZM48.1484 10.1445V21.5605H46.918V17.6094H45.1406C44.7031 18.8262 43.8896 19.5303 42.8438 19.5234C41.2441 19.5371 40.2051 17.8555 40.2051 15.2031C40.2051 12.5508 41.2441 10.8828 42.8438 10.8965C43.8896 10.8896 44.6963 11.5869 45.1406 12.7969H46.918V10.1445H48.1484ZM42.8438 12.1133C41.9414 12.1133 41.3945 13.248 41.3945 15.2031C41.3945 17.1582 41.9414 18.293 42.8438 18.293C43.7461 18.293 44.3203 17.1582 44.3203 15.2031C44.3203 13.248 43.7461 12.1133 42.8438 12.1133ZM45.4141 13.8496C45.4756 14.2666 45.5098 14.7178 45.5098 15.2031C45.5098 15.6816 45.4756 16.1328 45.4141 16.543H46.918V13.8496H45.4141ZM62.3809 9.91211V17.1992H61.082V14.0957H58.7988V13.0293H61.082V9.91211H62.3809ZM58.2793 17.541C60.8496 17.541 62.4219 18.3887 62.4219 19.8516C62.4219 21.3145 60.8496 22.1484 58.2793 22.1621C55.709 22.1484 54.123 21.3145 54.1367 19.8516C54.123 18.3887 55.709 17.541 58.2793 17.541ZM58.2793 18.5527C56.4883 18.5527 55.4219 19.0176 55.4219 19.8516C55.4219 20.6719 56.4883 21.1504 58.2793 21.1367C60.0703 21.1504 61.1367 20.6719 61.1367 19.8516C61.1367 19.0176 60.0703 18.5527 58.2793 18.5527ZM56.5566 12.0723C56.5566 13.6309 57.6641 15.1348 59.5098 15.7637L58.8398 16.7891C57.4658 16.3105 56.4541 15.333 55.9277 14.123C55.3877 15.4971 54.3145 16.5977 52.8516 17.1309L52.168 16.0918C54.0684 15.4355 55.2305 13.7812 55.2305 12.0859V11.8125H52.5781V10.7598H59.1816V11.8125H56.5566V12.0723Z" fill="#7E36E3"/>
-                                            </svg>
+                                            {/*scheduled.svg*/}
+                                            <img src="/icons/scheduled.svg" alt="이용예정" width={72} height={120} />
                                         </div>
                                         <div>
-                                            {/*개설확정.svg*/}
-                                            <svg width="93" height="32" viewBox="0 0 93 32" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                            <rect x="0.5" y="0.5" width="92" height="31" rx="15.5" fill="white"/>
-                                            <rect x="0.5" y="0.5" width="92" height="31" rx="15.5" stroke="#D3C1FF"/>
-                                            <path d="M16 15.2456L19.0877 18.3333L24.4209 13" stroke="#A065FF" strokeWidth="1.8" strokeLinecap="round" strokeLinejoin="round"/>
-                                            <path d="M42.9375 9.91211V22.1621H41.6934V15.8047H40.1758V21.5742H38.959V10.2402H40.1758V14.752H41.6934V9.91211H42.9375ZM37.7422 11.4707C37.7285 14.6973 36.6621 17.582 33.1621 19.5645L32.4102 18.6621C35.1309 17.124 36.2793 15.0391 36.4844 12.5371H32.9297V11.4707H37.7422ZM54.7637 9.91211V16.2148H53.4648V13.0703H50.8809V12.0176H53.4648V9.91211H54.7637ZM54.7637 16.7891V19.8379H48.0508V20.9727H55.1875V22.0117H46.7793V18.8535H53.4785V17.8145H46.752V16.7891H54.7637ZM48.8574 11.375C48.8574 12.9336 49.9785 14.3828 51.8242 14.9707L51.1406 15.9824C49.7939 15.5176 48.7754 14.5947 48.2422 13.4258C47.7021 14.7383 46.6494 15.7705 45.207 16.2695L44.5098 15.2305C46.4238 14.6152 47.5586 13.0977 47.5723 11.375V10.3086H48.8574V11.375ZM66.2891 9.92578V13.5625H68.0391V14.6289H66.2891V18.1699H64.9902V9.92578H66.2891ZM56.5684 16.8027C57.5186 16.8027 58.6191 16.8027 59.7676 16.7617V15.9414C58.373 15.7979 57.4844 15.1416 57.498 14.1504C57.4844 13.0293 58.6465 12.332 60.4102 12.3457C62.1602 12.332 63.3223 13.0293 63.3223 14.1504C63.3223 15.1348 62.4473 15.791 61.0664 15.9414V16.7207C62.1602 16.666 63.2676 16.5771 64.293 16.4336L64.4023 17.3633C61.7637 17.8281 58.8652 17.8418 56.7051 17.8418L56.5684 16.8027ZM66.2891 18.7441V22.1621H64.9902V19.7559H58.0996V18.7441H66.2891ZM63.9512 10.9648V11.9355H56.8828V10.9648H59.7676V9.81641H61.0664V10.9648H63.9512ZM60.4102 13.2344C59.3574 13.2207 58.7285 13.5625 58.7285 14.1504C58.7285 14.7246 59.3574 15.0938 60.4102 15.0938C61.4492 15.0938 62.0781 14.7246 62.0781 14.1504C62.0781 13.5625 61.4492 13.2207 60.4102 13.2344ZM78.9629 9.91211V17.1992H77.6641V14.0957H75.3809V13.0293H77.6641V9.91211H78.9629ZM74.8613 17.541C77.4316 17.541 79.0039 18.3887 79.0039 19.8516C79.0039 21.3145 77.4316 22.1484 74.8613 22.1621C72.291 22.1484 70.7051 21.3145 70.7188 19.8516C70.7051 18.3887 72.291 17.541 74.8613 17.541ZM74.8613 18.5527C73.0703 18.5527 72.0039 19.0176 72.0039 19.8516C72.0039 20.6719 73.0703 21.1504 74.8613 21.1367C76.6523 21.1504 77.7188 20.6719 77.7188 19.8516C77.7188 19.0176 76.6523 18.5527 74.8613 18.5527ZM73.1387 12.0723C73.1387 13.6309 74.2461 15.1348 76.0918 15.7637L75.4219 16.7891C74.0479 16.3105 73.0361 15.333 72.5098 14.123C71.9697 15.4971 70.8965 16.5977 69.4336 17.1309L68.75 16.0918C70.6504 15.4355 71.8125 13.7812 71.8125 12.0859V11.8125H69.1602V10.7598H75.7637V11.8125H73.1387V12.0723Z" fill="#A065FF"/>
-                                            </svg>
+                                            {/*sonfirmed.svg*/}
+                                            <img src="/icons/confirmed.svg" alt="개설확정" width={90} height={120} />
                                         </div>
                                     </div>
                                         <h3 className="py-2">달램핏 오피스 스트레칭 | <span>을지로 3가</span> </h3>
