+++ vercel.json
@@ -1,9 +1,9 @@
 {
-    "git": {
-        "deploymentEnabled": {
-            "staging": true
-        }
-    },
-    "framework": "nextjs",
-    "outputDirectory": ".next"
+    "eslint.enable": true,
+    "eslint.validate": [
+        "javascript",
+        "javascriptreact",
+        "typescript",
+        "typescriptreact"
+    ]
 }
